DataTypes

Description

Module with all the datatypes definitions needed for the Pacman

Documentation

type Maze = [Corridor]

type Corridor = [Piece]

data Piece

Constructors

Food FoodType

PacPlayer Player

Empty

Wall

data FoodType

Represents the different kinds of food

Constructors

Big
Rarer, worth more points, Pacman enters Mega mode when it eats it

Little
More common, worth less, Pacman does not change mode when it eats it

data Play

Represents an action made by a player Each action is a movement of a player in a certain direction

Warning

An action with a certain orientation does not mean that the player will move in that direction, only that it will be facing in that direction after the action

Constructors

Move

Fields

Int

The id of the player who performs the action

Orientation

The orientation of the action

data Orientation

Represents directions: Left Right Up / Down

Constructors

L
Left

R
Right

U
Up

D
Down

type Coords = (Int, Int)

Coordinates in the maze The maze can be as a two-dimensional grid, where the origin (0,0) is the top left corner. The x-axis is the horizontal axis and the y-axis the vertical axis. The y coordinate increases as you go down, and the x coordinate increases as you go to the right

compareCoords

Arguments

:: Coords
The first set of coordinates

-> Coords
The second set of coordinates

-> Bool
Whether they are equal or not

Tests wether two coordinates are the same Eq is a class, only applies to data types. Coords is a tuple

Examples

>>> compareCoords (1,2) (1,2)
True

>>> compareCoords (1,3) (3,1)
False

data State

Represents the state of the Game at a given time

Constructors

State

Fields

maze :: Maze

The maze of the game

playersState :: [Player]

The players in the game

level :: Int

The current level of the game

data Player

Represents a player. Each player can be uniquely described by a PacState or a GhoState depending if it is a pacman or a ghost

Constructors

Pacman PacState
Player is a pacman, described by its PacState

Ghost GhoState
Player is a ghost, described by its GhoState

data PacState

Represents the state of Pacman in a given point in time

Example

pac = PacState (0,(3,2),1,R,5,1) 0 Open Normal

Constructors

PacState

Fields

pacState :: PlayerState

timeMega :: Double

openClosed :: Mouth

pacmanMode :: PacMode

data GhoState

Represents the state of a ghost in a given point in time

Example

ghost = GhoState (0,(3,2),1,R,5,1) Alive

Constructors

GhoState

Fields

ghostState :: PlayerState

ghostMode :: GhostMode

type PlayerState = (Int, Coords, Double, Orientation, Int, Int)

Represents the state of a generic player (i.e. Pacman or ghosts) in a given point in time

data Mouth

Represents the state of Pacman's mouth (open or closed)

Constructors

Open

Closed

data PacMode

Represents the state of Pacman (Dying Normal or Mega)

Constructors

Dying

Mega

Normal

data GhostMode

Represents the state of a ghost (dead or alive)

Constructors

Dead

Alive

type Instructions = [Instruction]

Type used to compact the entire maze

data Instruction

Data type used to compact a corridor (horizontal)

Constructors

Instruct [(Int, Piece)]
There is the same piece in n consecutive squares

Repeat Int
The same as the instruction with index i