Game

Description

Module responsible for handling the events (keyPresses, passing of time) and the drawing of everything related to the gameplay (the maze, the players, score, lives, .etc)

Documentation

drawPauseScreen

Arguments

:: PictureSet
The given set of pictures

-> [Picture]
The pictures corresponding to the pause screen

Given the set of pictures, returns the pictures corresponding to the pause screen

drawLostScreen

Arguments

:: PictureSet
The given set of pictures

-> [Picture]
The pictures corresponding to the death screen

Given the set of pictures, returns the pictures corresponding to the death screen

squareSize :: Float

Default size of maze square (30px x 30px)

orientationID

Arguments

:: Orientation
The given orientation

-> Int
The orientation's id

Given an orientation, returns its unique ID. The ID corresponds, not only to the position in the list of pictures regarding the imagery for the ghosts, but also to the number of times one has to rotate the image of the player facing to the right by 90 degrees in order for it to be in the correct orientation.

rotatePicture

Arguments

:: Picture
The given picture (of a player pointing to the *right*)

-> Orientation
The player's orientation

-> Picture
The resulting picture

Given a picture corresponding to a player, and its orientation, returns the picture resulting from rotating the original to the correct position.

drawPieces

Arguments

:: PictureSet
The given set of pictures

-> Player
The controlled player

-> Maze
The given maze

-> (Float, Float)
The coordinates of the top left corner of the maze

-> (Float, Float)
The coordinates to render the current picture on

-> [Picture]
The resulting set of pictures

Given a set of pictures, a maze, the coordinates to place the current picture, and the coordinates of the top left corner of the maze, returns the set of pictures corresponding to the rendering of the maze

repeatPicture

Arguments

:: Int
The number of times to repeat the picture

-> Picture
The given picture

-> (Float, Float)
The coordinates to place the first picture in

-> [Picture]
The resulting set of pictures

Given a number x, a Picture, and a point, returns the list of pictures resulting from repeating the given picture x times in the horizontal axis.

It is an auxiliary function to drawPacLives

drawPacLives

Arguments

:: PictureSet
The given set of pictures

-> Player
The given player

-> Picture
The resulting picture, containing the pictures corresponding to the lives of the pacman

Given the set of images and a player (pacman), returns the picture corresponding to the number of lives remaining.

drawMega

Arguments

:: Player
The given player

-> Picture
The picture of the mega mode

Given a pacman, returns the picture corresponding to the mega mode of the pacman.

This corresponds to a vertical bar on the top of the window which becomes shorter and shorter until the pacman runs out of mega mode.

drawLevel

Arguments

:: Int
The given level

-> Picture
The picture corresponding to the given level

Given the current level, returns the picture corresponding to said level

drawScore

Arguments

:: Player
The given score

-> Picture
The picture corresponding to the given score

Given the current score, returns the picture corresponding to said score

drawGame

Arguments

:: PictureSet
The given set of pictures

-> Manager
The current game state

-> Picture
The resulting picture

Given a set of pictures, and the current game state, returns the picture corresponding to the rendering of the game state

drawLevelUp

Arguments

:: PictureSet
The given set of pictures

-> Picture
The picture for the level up display

Given the set of pictures, draws the picture corresponding to the level up screen

nextGhost

Arguments

:: [Player]
The given list of players

-> Int
The id of the currently controlled ghost

-> Int
The id of the next ghost

Given a list of players, and the id of the ghost being controlled, returns the id of the next ghost.

reactEventGame

Arguments

:: Event
The given event

-> GlossState
The given state

-> GlossState
The resulting state

Given an event (key press) and a state, returns the state after the event is processed

reactTimeGame

Arguments

:: Float
The time elapsed since the last frame

-> GlossState
The previous state

-> GlossState
The current state

Given the time elapsed since the last frame was displayed, and the previous state, returns the state after considering the passing of time

resetTimer

Arguments

:: Manager
The given manager

-> Manager
The resulting manager

Resets the delta and before to zero and the current time, respectively