LazyBFS

Description

Module used for calculating the distances between squares of the maze. Used by Tarefa5 to compute the moves the ghosts should make.

Documentation

maxValue :: Int

The maximum value of an integer (is returned when there is no path from A to B)

type Memoization = Map Coords ([Set Coords], Map Coords (Maybe Int))

Stores the distance from every point to every other point

distance

Arguments

:: Memoization
the given mamoization

-> Coords
the first point

-> Coords
the second point

-> Int
the distance between them, or maxValue if there is no path from one to the other

Given a Memoization, extracts from it the distance between 2 points

mazeSize

Arguments

:: Memoization
the given Mamoization object

-> Coords
the size of the original maze (height, width)

Given a Memoization, returns the size of the original maze

getMemory

Arguments

:: Maze
the given maze

-> Memoization
the resulting Memoization object

Given a maze, creates a Memoization object storing the distance between every 2 points

mapper

Arguments

:: Maze
the maze

-> Coords
the key

-> a
the original value (unused)

-> ([Set Coords], Map Coords (Maybe Int))
the resulting data structure

Populates an empty entry with the intended data structure using the key

coords

Arguments

:: Maze
the original maze (used for size)

-> Coords
the given point

-> Map Coords (Maybe Int)
The map of points

Returns a Map containing every point bigger than the given point. Used to ensure no 2 pairs of points in the final map are equivalent.

atDists

Arguments

:: Maze
the given maze

-> Coords
the source point

-> [Set Coords]
the list of sets

Given a maze and a source point, returns a list of Sets, where the n-th Set contains all the points of the maze distanced n from the source

atDists'

Arguments

:: Maze
the given maze

-> Coords
the source point

-> Set Coords
the set before the previous

-> Set Coords
the previous set

-> [Set Coords]
the list of sets

Auxiliar function to atDists. Recursively creates the list of Sets from the 2 previous sets

getNeighbours

Arguments

:: Maze
the given maze

-> Coords
the source point

-> [Coords]
the adjacent points

Given a maze and a source point, returns the adjacent points to the source that are clear (for definition of "clear", check PacmanUtils.isSquareClear)