Main

Description

Main module for the Gloss implementation of the game. It handles the window creation, redirects the drawing and reaction to events to the correct modules, and loads the PictureSet used throughout the program's lifespan.

Usage of pictures

Pictures were used, many times in place of text, due to the limitations of gloss when it comes to text rendering, namely custom fonts.

Documentation

background :: Color

The default background color

initialState

Arguments

:: PictureSet
The given set of pictures

-> GlossState
The resulting initial state

The default initial state for gloss rendering

drawState

Arguments

:: PictureSet
The given set of pictures

-> GlossState
The current state

-> Picture
The picture to be rendered on screen

Given the set of pictures and the current state, returns the picture to be rendered on screen

reactEvent

Arguments

:: Event
The given event

-> GlossState
The current state

-> GlossState
The resulting state

Given an event (key press) and a state, returns the state after the event is processed

reactTime

Arguments

:: Float
The time elapsed since the last frame

-> GlossState
The previous state

-> GlossState
The current state

Given the time elapsed since the last frame was d Int -- ^ Pacman's lives

fr :: Int

Default frame rate (60)

dm :: Display

Default display

loadGhostAlivePictures

Arguments

:: IO ([Picture], [Picture], [Picture])
([Red], [Pink], [Green])

Returns the pictures to be used for rendering of the alive ghosts.

They are divided by color: first are the red ones, then the pink ones, and finally the green ones.

loadGhostDeadPictures

Arguments

:: IO ([Picture], [Picture])
([Dark, Light])

Returns the pictures to be used for rendering of the dead ghosts.

They are divided by colors : first the dark blue ones, then the light blue ones.

loadMazePicture

Arguments

:: FilePath
The file containing the maze

-> IO Picture
The resulting picture

Loads the pictures to be displayed when choosing the given maze

loadMazePicturesAux

Arguments

:: [FilePath]
The list of files containing mazes

-> IO [Picture]
The resulting pictures

Auxiliary function to loadMazePicture

loadMazePictures :: IO [(String, Picture)]

Returns the pictures displayed when choosing a maze

loadPictureSet :: IO PictureSet

Returns the set of pictures to be used throughout the execution of the program (player images, for example).

main :: IO ()

Starting point of the application