MazeMenu

Description

Module responsible for handling the UI of the Maze Menu and the corresponding inputs.

Documentation

drawMazeMenu

Arguments

:: Int
The maze pressed

-> PictureSet
The PictureSet of the application

-> Picture
The resulting picture

Returns the Picture used to render the Maze Menu

reactEventMazeMenu

Arguments

:: Event
The given event

-> GlossState
The current state

-> GlossState
The resulting state

Given an event (key press) and a state, returns the state after the event is processed