PacmanUtils

Description

Module containing generic auxiliary functions required by multiple modules.

Documentation

testMaze :: Maze

Variable used to test functions

testMaze = [
    [Wall, Wall, Wall, Wall, Wall, Wall, Wall, Wall],
    [Empty, Food Big, Food Little, Food Little, Food Little, Empty, Food Little, Empty],
    [Empty, Food Little, Food Little, Empty, Food Little, Wall, Food Little, Empty],
    [Wall, Wall, Wall, Wall, Wall, Wall, Wall, Wall]
       ]

js :: [Player]

Variable used to test functions

js = [jogador1, jogador2]

jogador1 :: Player

Variable used to test functions

jogador1= (Pacman (PacState (0,(2,3),1,R,5,1) 0 Open Normal))

jogador2 :: Player

Variable used to test functions

jogador2 = (Ghost (GhoState (1,(1,5),1,L,10,0) Alive))

getSize

Arguments

:: Maze
the given maze

-> Coords
the (y, x) size of the maze

Returns the (y, x) size of the given maze, where y is vertical and x is horizontal

Example

>>> getSize testMaze
(8, 4)

getPieceFrom

Arguments

:: Maze
The coordinates of the square

-> Coords
The given maze

-> Piece
The piece found at the given coordinates

Returns the Piece at the specified coordinates. For out of bounds coordinates an error is thrown

This is a recursive function which first iterates to the correct corridor (y coordinate), and, from there, locates the correct piece on that corridor (x coordinate)

Note that (y, x) corresponds to the distance vertically and horizontally, respectively, to the wanted square (origin is (0,0))

Examples

>>> getPieceFrom (1,1) testMaze
Food Big

>>> getPieceFrom (3,3) testMaze
Wall

isSquareClear

Arguments

:: Coords
The coordinates of the square

-> Maze
The given maze

-> Bool
Whether the square is clear or not

Checks if square of given coordinates is clear, i.e, if it is not a wall.

This function uses the function getPieceFrom to get the piece and then compare it with a Wall

Note that (y, x) corresponds to the distance vertically and horizontally, respectively, to the wanted square (origin is (0,0))

Examples

>>> isSquareClear (1,1) testMaze
True

>>> isSquareClear (3,3) testMaze
False

replaceSquareWith1D

Arguments

:: Corridor
the corridor to change

-> Int
index of the value to replace

-> Piece
piece to substitute the original one with

-> Corridor
the resulting corridor

Replaces the x-th value of a corridor with the given piece. For out of bounds coordinates an error is thrown. Used as an auxiliary function to replaceSquareWith

Example

>>> replaceSquareWith1D [Wall, Wall, Wall, Wall, Wall, Wall, Wall, Wall] 4 (Food Little)
[#,#,#,#,.,#,#,#]

replaceSquareWith

Arguments

:: Coords
coordinates to change

-> Piece
piece to substitute original one with

-> Maze
the maze to change

-> Maze
the resulting maze

Replaces the value (y, x) of the maze with the given piece. For out of bounds coordinates an error is thrown

This is a recursive function which first iterates to the correct corridor (y coordinate), and, from there, locates the correct piece on that corridor (x coordinate)

Note that (y, x) corresponds to the distance vertically and horizontally, respectively, to the wanted square (origin is (0,0))

Example

>>> replaceSquareWith testMaze (2,2) Wall
[[#,#,#,#,#,#,#,#],[ ,o,.,.,.,.,., ],[ ,.,#,.,.,#,., ],[#,#,#,#,#,#,#,#]]

firstFood

Arguments

:: Maze
The given maze

-> Maybe Coords
The coordinates of the first food (Nothing if there is no food)

Given a maze, returns the coordinates of the first food, big or small, or Nothing if the isn't any

The food is searched up to down, left to right (similar to reading)

isGameOver

Arguments

:: State
The given state

-> Bool
Whether there is food on the maze or not

Given a state, returns whether or not the game is over by the food remaining, i.e, if there is no food on the maze, then the game is over.

isPacman

Arguments

:: Player
The given player

-> Bool
Whether or not the given player is a pacman

Checks if the given player is a pacman

Examples

>>> isPacman (Pacman (PacState (1,(1,1),1.0,U,1,1) 0.0 Open Normal))
>>> True

>>> isPacman (Ghost (GhoState (1,(1,1),1.0,U,1,1) Alive))
>>> False

hasOrientation

Arguments

:: Player
The given player

-> Orientation
The orientation we want to compare the player's with

-> Bool
Whether the orientation is the same or not

Checks if a player has a certain orientation

Examples

>>> hasOrientation jogador1 R
True

>>> hasOrientation jogador2 D
False

rotatePlayerTo

Arguments

:: Orientation
The new orientation

-> Player
The given player

-> Player
The resulting player

Rotates the given player to the specified orientation

Example

>>> rotatePlayerTo U jogador1
Pacman (PacState {pacState = (0,(3,2),1.0,U,5,1), timeMega = 0.0, openClosed = Open, pacmanMode = Normal})

setPlayerCoordinates

Arguments

:: Coords
The new coordinates

-> Player
The given player

-> Player
The resulting player

Sets the coordinates of the player

Example

>>> setPlayerCoordinates (1,1) jogador1
Pacman (PacState {pacState = (0,(1,1),1.0,R,5,1), timeMega = 0.0, openClosed = Open, pacmanMode = Normal})

updatePlayerScore

Arguments

:: Int
the score change. Positive to increase the score, negative to decrease the score

-> Player
the given player

-> Player
the resulting player (i.e. the player with updated score)

Changes the score of the given player by the specified amount.

Examples

>>> updatePlayerScore 20 jogador1
Pacman (PacState {pacState = (0,(3,2),1.0,R,25,1), timeMega = 0.0, openClosed = Open, pacmanMode = Normal})

>>> updatePlayerScore (-10) jogador1
Pacman (PacState {pacState = (0,(3,2),1.0,R,-5,1), timeMega = 0.0, openClosed = Open, pacmanMode = Normal})

setPacmanMode

Arguments

:: PacMode
The new PacMode

-> Player
The original player

-> Player
The resulting player (i.e. the original player with the new PacMode)

Updates given player's pacmode to the specified mode. For ghosts an error is thrown

Example

>>> setPacmanMode Mega jogador1
Pacman (PacState {pacState = (0,(3,2),1.0,R,5,1), timeMega = 0.0, openClosed = Open, pacmanMode = Mega})

getTimeMega

Arguments

:: Player
The given player

-> Double
The player's time mega

Returns the given player's timeMega. For ghosts an error is thrown

Example

>>> getTimeMega jogador1
0

setTimeMega

Arguments

:: Double
The given value

-> Player
The original player

-> Player
The resulting player

Sets the given player's timeMega. For ghosts an error is thrown

Example

>>> setTimeMega 1.0 jogador1
Pacman (PacState {pacState = (0,(3,2),1.0,R,5,1), timeMega = 1.0, openClosed = Open, pacmanMode = Normal})

nextTimeMega

Arguments

:: Player
The given player

-> Double
The value to which timeMega should be updated

Returns the value to which timeMega should be updated during the next call of play on that player. This value is calculated by subtracting from the current value the time elapsed since the last call of play (on average). This function is (unsurprisingly) an auxiliary function to play.

getPlayerSpeed

Arguments

:: Player
The given player

-> Double
The player's speed

Returns the given player's speed

Example

>>> getPlayerSpeed jogador1
1

setPlayerSpeed

Arguments

:: Double
The new speed

-> Player
The player to update the speed

-> Player
The resulting player

Sets the player's speed

Example

>>> setPlayerSpeed 2.27 jogador1
Pacman (PacState {pacState = (0,(3,2),2.27,R,5,1), timeMega = 0.0, openClosed = Open, pacmanMode = Normal})

getLives

Arguments

:: Player
The given player

-> Int
The number of lives of the player

Returns the number of lives of the given player

Example

>>> getLives jogador1
1

stealLife

Arguments

:: Player
The given player

-> Player
The given player with one less life

Steals one life from the specified player (i.e. life = life - 1)

Example

>>> stealLife jogador1
Pacman (PacState {pacState = (0,(3,2),1.0,R,5,0), timeMega = 0.0, openClosed = Open, pacmanMode = Normal})

toggleMouth

Arguments

:: Player
The given player

-> Player
The resulting player

Toggles the mouth state for pacman (i.e. if it is closed, sets it to open and vice-versa). For ghosts an error is thrown

Example

>>> toggleMouth jogador1
Pacman (PacState {pacState = (0,(3,2),1.0,R,5,1), timeMega = 0.0, openClosed = Closed, pacmanMode = Normal})

setMouth

Arguments

:: Mouth
The specified mouth state

-> Player
The original player

-> Player
The resulting player

Sets the player's mouth to the specified state. For ghosts an error is thrown

Example

>>> setMouth Closed jogador1
Pacman (PacState {pacState = (0,(3,2),1.0,R,5,1), timeMega = 0.0, openClosed = Closed, pacmanMode = Normal})

getPlayer

Arguments

:: [Player]
List of players to search

-> Int
Id of wanted player

-> Player
The wanted player

Returns the player whose id is equal to x. If no such player exists, an error is thrown

Examples

>>> getPlayer js 0
Pacman (PacState {pacState = (1,(6,15),1.0,L,10,0), timeMega = 0.0, openClosed = Closed, pacmanMode = Normal})

>>> getPlayer js 2
*** Exception: Player not found

getPacman

Arguments

:: [Player]
The list of players

-> Player
The first pacman

Retrieves the first pacman player from the given list of players. If no pacman is found, an error is thrown

areTheSame

Arguments

:: Player
The first player

-> Player
The second player

-> Bool
Whether they are equal (i.e. have the same id and type)

Tests whether or not two players are the same by comparing their type and id

Examples

>>> areTheSame jogador1 jogador1
True

>>> areTheSame jogador1 jogador2
False

updatePlayerList

Arguments

:: Player
The new player

-> [Player]
The initial list

-> [Player]
The resulting list

Updates the list by replacing the equal player in the list with the specified player. If no such player is found an error is thrown

It is an auxiliar function used by updatePlayer. Comparisons are done via the areTheSame function

Example

>>> updatePlayerList (Pacman (PacState (0,(3,2),1,U,5,1) 0 Open Normal)) js
[Pacman (PacState {pacState = (0,(3,2),1.0,U,5,1), timeMega = 0.0, openClosed = Open, pacmanMode = Normal}),Ghost (GhoState {ghostState = (1,(6,1),1.0,L,10,0), ghostMode = Alive})]

updatePlayer

Arguments

:: Player
The new player

-> State
The original state

-> State
The resulting state

Updates the state by replacing the equal player in the state with the specified player. The replacement is done in the player list (by direct replacement)

Example

>>> updatePlayer (Pacman (PacState (0,(3,2),1,U,5,1) 0 Open Normal)) state
State {maze = [[#,#,#,#,#,#,#,#],[ ,*,.,.,.,M,., ],[ ,.,.,u,.,#,., ],[#,#,#,#,#,#,#,#]], playersState = [Pacman (PacState {pacState = (0,(3,2),1.0,U,5,1),
timeMega = 0.0, openClosed = Open, pacmanMode = Normal}),Ghost (GhoState {ghostState = (1,(6,1),1.0,L,10,0), ghostMode = Alive})], level = 1}