PlayMenu

Description

Module responsible for handling the UI of the Play Menu and the corresponding inputs.

Documentation

drawPlayMenu

Arguments

:: PictureSet
The given PictureSet

-> Int
The button pressed

-> Picture
The resulting picture

Returns the Picture used to render the Play Menu

title

Arguments

:: PictureSet
The given PictureSet

-> Picture
The resulting Picture

the title of the page

pacmanButton

Arguments

:: PictureSet
The given PictureSet

-> Bool
Whether or not the button is selected

-> Picture
The picture corresponding to the button

The top button (the one to choose to play as pacman)

ghostButton

Arguments

:: PictureSet
The given PictureSet

-> Bool
Whether or not the button is selected

-> Picture
The picture corresponding to the button

The middle button (the one to choose to play as the ghosts)

backButton

Arguments

:: PictureSet
The given PictureSet

-> Bool
Whether or not the button is selected

-> Picture
The picture corresponding to the button

The lowest button (to go back in the menus)

choosePacman

Arguments

:: Manager
The given manager

-> Manager
The resulting manager

chooseGhost

Arguments

:: Manager
The given manager

-> Manager
The resulting manager

reactEventPlayMenu

Arguments

:: Event
The given event

-> GlossState
The current state

-> GlossState
The resulting state

Given an event (key press) and a state, returns the state after the event is processed