Tarefa1

Description

Module which generates valid mazes as per the task statement.

A maze is considered valid if and only if it is at least of size 15x10, has a tunnel connecting the two ends of it in the middle row (or middle rows, if of even height), and has the ghost house placed in the middle of it.

Documentation

generateRandoms

Arguments

:: Int
The length of the wanted list

-> Int
The given seed

-> [Int]
The resulting list of numbers

Generates a list of n pseudo-randoms integers based on the given seed

Examples

>>> generateRandoms 5 123
[43,5,7,30,50]

>>> generateRandoms 7 4124354
[89,72,69,7,20,33,49]

convertIntListToCorridor

Arguments

:: [Int]
List of pseudo-random ints

-> Corridor
Corresponding corridor

Given a list of pseudo-random integers, returns a Corridor based on the probabilities of each type of Piece

Examples

>>> convertIntListToCorridor [3,45,98,1,0,4]
[o,.,#,.,.,.]

>>> convertIntListToCorridor [56,9,12,9,75,90,34]
[.,.,.,.,#,#,.]

generateWall

Arguments

:: Int
The length of the desired corridor

-> Corridor
Resulting corridor

Returns a corridor of length len consisting of only Wall

Examples

>>> generateWall 5
[#,#,#,#,#]

getHouseLength

Arguments

:: Int
The width of the maze

-> Int
The width of the ghostHouse

Returns the length of the ghost house including walls (8 if length is even, 9 otherwise)

Examples

>>> getHouseLength 20
>>> 8

>>> getHouseLength 31
>>> 9

generateArroundHouse

Arguments

:: Int
The width of the maze

-> Corridor
The original corridor

-> Corridor
The new corridor

Fixes the corridors directly above and below the house (see restrictions)

Example

>>> generateArroundHouse 20 (generateWall 20)
>>> [#,#,#,#,#, , , , , , , , , , ,#,#,#,#,#]

generateInsideHouse

Arguments

:: Int
The width of the maze

-> Corridor
The original corridor

-> Corridor
The new corridor

Fixes the Corridor that passes through the ghost house

It also makes sure that the tunnel is clear, i.e, that it is possible to reach the ghost house from the tunnel. It does this by replacing all walls from the generated maze with Empty

Example

>>> generateInsideHouse 20 (generateWall 20)
>>> [ , , , , , ,#, , , , , , ,#, , , , , , ]

generateUpperHouse

Arguments

:: Int
The width of the maze

-> Corridor
The original corridor

-> Corridor
The new corridor

Generates the Corridor containing the door to the ghost house

Example

>>> generateUpperHouse 20 (generateWall 20)
>>> [#,#,#,#,#, ,#,#,#, , ,#,#,#, ,#,#,#,#,#]

generateBottomHouse

Arguments

:: Int
The width of the maze

-> Int
The height of the maze

-> Corridor
The original corridor

-> Corridor
The new corridor

Generates the Corridor containing the bottom of the ghost house.

Requires both width and height of the maze

Example

>>> generateBottomHouse 20 10 (generateWall 20)
>>> [ ,#,#,#,#, ,#,#,#,#,#,#,#,#, ,#,#,#,#, ]

generateStartingMaze

Arguments

:: Int
The width of the maze

-> Int
The height of the maze

-> Maze
The resulting maze

Generates a Maze consisting of only walls

Example

>>> generateStartingMaze 5 2
>>> [[#,#,#,#,#], [#,#,#,#,#]]

changeVal1D

Arguments

:: [a]
The original list

-> Int
Position of the element to change

-> a
The new element

-> [a]
The resulting list

Given a list, returns the list obtained by replacing the x-th element (base 0) with the given element

Example

>>> changeVal1D [1,2,3] 1 0
>>> [1,0,3]

changeVal

Arguments

:: [[a]]
The original list of lists

-> (Int, Int)
The coordinates (x,y) (base 0) of the value to replace

-> a
The new value

-> [[a]]
The resulting list

Given a list of lists, returns the list obtained by replacing the element at position (x,y), i.e, the x-th element of the y-th list (indices base 0), with the given element

Example

>>> changeVal [[1,2], [3,4]] (1,1) 0
>>> [[1,2], [3,0]]

get

Arguments

:: [[a]]
The list of lists

-> (Int, Int)
The coordinates (x,y) (base 0) of the wanted element

-> a
The wanted element

Given a list of lists, returns the element at position (x,y), i.e, the x-th element of the y-th list (indices base 0)

Example

>>> get [[1,2], [3,4]] (1,1)
4

first

Arguments

:: (a, b, c)
The given triple

-> a
The first element

Returns the first element of a triple

Example

>>> first (1,2,3)
>>> 1

selectSquare

Arguments

:: Maze
The given maze

-> (Int, Int)
The coordinates of the given square

-> (Int, Int)
The size of the maze (width, height)

-> [[Bool]]
The list of visited squares

-> [Piece]
The list of pieces to choose from

-> (Maybe Piece, [Piece])
The resulting piece and leftover pieces

Given a Maze, the coordinates of the current square, the size of the maze, the list of visited squares and the list of pieces to choose from, returns the piece to place on that square (if square out of bounds or already visited, returns Nothing, otherwise, 'Just Piece'), as well as the pieces with have not been used, so they can be selected in the future

Examples

>>> used = take 20 (repeat (take 20 (repeat False)))
>>> selectSquare (generateStartingMaze 20 20) (5,5) (20,20) used [Wall, Empty, Wall, Food Little]
>>> (Just #,[ ,#,.])

truncate'

Arguments

:: (Maybe Piece, Maybe Piece, Maybe Piece, Maybe Piece)
The original set of pieces

-> (Maybe Piece, Maybe Piece, Maybe Piece, Maybe Piece)
The new set of pieces, containing at least one Food Little

Given a 4-uple of potential pieces, returns a valid 4-uple, i.e., returns a 4-uple in where at least one of the component is either a Food or Empty.

This makes sure that the DFS will continue until all squares have been visited and that it will not stop because it decided to surround itself in walls.

Examples

>>> truncate' (Nothing, Nothing, Nothing, Nothing) = (Nothing, Nothing, Nothing, Nothing)
>>> truncate' (Just Wall, Just Wall, Just Wall, Nothing) = (Just (Food Little), Just (Food Little), Just (Food Little), Nothing)
>>> truncate' (Just Wall, Just Wall, Just Wall, Just Wall) = (Just (Food Little), Just (Food Little), Just (Food Little), Just (Food Little))
>>> truncate' (Just Empty, Just Empty, Just Empty, Just (Food Little)) = (Just Empty, Just Empty, Just Empty, Just (Food Little))

selectMove

Arguments

:: Maze
The given maze

-> (Int, Int)
The coordinates (x,y) of the square

-> (Int, Int)
The size of the maze (width, height)

-> [[Bool]]
The list of visited squares

-> [Piece]
The list of pieces to choose from

-> ((Maybe Piece, Maybe Piece, Maybe Piece, Maybe Piece), [Piece])
The resulting moves (if not possible to move to neighbour, Nothing, otherwise, 'Just Piece') and leftover pieces

Given a maze, a position in said maze, a list of visited squares and a list of pieces to choose from, returns the pieces to replace all neighbours of the current square with

The choice for each component is done by the selectSquare function

Examples

>>> used = take 20 (repeat (take 20 (repeat False)))
>>> selectMove (generateStartingMaze 20 20) (3,3) (20,20) used [Wall, Empty, Wall, Food Little]
>>> ((Just .,Just #,Just  ,Just #),[])

dfs

Arguments

:: Maze
The starting maze (at first only containing walls)

-> (Int, Int)
The coordinates (x,y) of the current square

-> (Int, Int)
The width and height of the maze

-> [[Bool]]
The list of which squares have been visited (at first all values are False)

-> [Piece]
The list of pieces to choose from

-> Piece
The piece to replace current square with

-> (Maze, [[Bool]], [Piece])
The resulting maze, the list of used squares and the leftover pieces for future choices

Given a starting maze consisting of only walls, a list of pieces and a piece, constructs a maze where every square is reachable from any other square (excluding walls, for obvious reasons)

As the name implies, this function transverses the maze depth-first. When it reaches a square, if it has not been visited before and is within bounds, replace the wall with the value of the piece argument and mark it as visited. It then send the list of pieces to the selectMove function, which will select the pieces to place on the other neighbouring squares which haven't been visited yet.

After this, it visits each of the neighbouring squares and replaces their Wall with the selected piece

Example

>>> maze = generateStartingMaze 20 20
>>> used = take 20 (repeat (take 20 (repeat False)))
>>> pieces = convertIntListToCorridor (generateRandoms 600 123)
>>> dfs maze (1,1) (20,20) used pieces (Food Little)
>>> ([[#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#],
      [#,.,.,.,#,.,.,.,.,#,#,#,.,.,.,.,.,#,.,#],
      [#,.,.,.,.,.,.,#,.,.,.,.,.,#,.,.,.,.,.,#],
      [#,.,.,#,.,.,.,.,.,.,.,.,.,.,.,#,#,.,.,#],
      [#,#,.,.,.,#,#,.,.,#,.,.,.,.,.,.,.,.,.,#],
      [#,.,#,#,.,.,.,.,.,.,.,.,.,.,.,.,.,#,.,#],
      [#,.,.,.,.,.,#,.,.,.,.,#,.,.,#,.,.,.,.,#],
      [#,.,.,.,.,.,#,#,.,.,.,.,#,.,.,#,#,.,.,#],
      [#,.,.,.,.,.,.,.,#,.,.,.,#,#,.,.,.,.,.,#],
      [#,.,#,.,#,.,.,.,.,.,.,.,.,.,.,.,.,.,.,#],
      [#,.,.,#,#,#,.,#,.,#,.,.,.,.,.,.,.,.,.,#],
      [#,#,.,#,#,.,.,#,.,.,.,.,.,.,.,.,#,o,.,#],
      [#,.,.,.,.,.,.,.,.,.,.,.,.,.,.,.,.,.,.,#],
      [#,.,#,.,.,.,.,.,.,.,#,.,.,.,.,#,.,.,.,#],
      [#,.,.,.,.,.,.,#,#,.,.,.,.,.,.,.,.,.,.,#],
      [#,.,.,.,.,.,.,#,#,.,.,#,#,.,.,.,.,.,.,#],
      [#,.,#,.,.,.,.,.,.,.,.,.,.,#,.,#,.,.,.,#],
      [#,.,.,.,.,.,.,.,.,.,.,.,.,.,.,.,.,.,.,#],
      [#,.,.,#,#,.,.,#,.,.,.,.,.,.,.,#,.,.,.,#],
      [#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#]],
     [[False,False,False,False,False,False,False,False,False,False,False,False,False,False,False,False,False,False,False,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,True,False],
      [False,False,False,False,False,False,False,False,False,False,False,False,False,False,False,False,False,False,False,False]],
     [.,#,.,#,.,.,.,.,.,.,#,.,.,.,.,#,#,.,.,#,.,.,.,.,.,#,.,.,.,.,.,#,.,.,.,.,.,.,#,#,#,.,#,.,.,#,#,.,#,#,.,.,.,.,#,#,.,.,.,#,.,#,.,.,.,#,.,#,#,.,#,.,#,.,.,.,.,.,.,#,.,#,.,.,#,#,.,#,.,.,.,.,o,#,.,.,.,#])

fixMaze

Arguments

:: (Int, Int)
Width and height of the maze

-> Int
Index of current corridor (base 1)

-> Maze
The original maze without structures

-> Maze
The resulting maze with structures

Given a maze surrounded by walls without any structures, returns the given maze with the ghost house and tunnel inserted

It iterates through each corridor, and based on the index of the current corridor, selects the appropriate structure to add

Example

>>> fixMaze (20, 7) 1 (generateStartingMaze 20 7)
>>> [[#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#],
     [#,#,#,#,#, , , , , , , , , , ,#,#,#,#,#],
     [#,#,#,#,#, ,#,#,#, , ,#,#,#, ,#,#,#,#,#],
     [ ,.,.,.,., ,#, , , , , , ,#, ,.,.,.,., ],
     [#,#,#,#,#, ,#,#,#,#,#,#,#,#, ,#,#,#,#,#],
     [#,#,#,#,#, , , , , , , , , , ,#,#,#,#,#],
     [#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#,#]]

generateMaze

Arguments

:: Int
Width of the maze

-> Int
Height of the maze

-> Int
The seed used to generate random numbers

-> Maze
The resulting maze

Generates a maze of the given size based on the given seed

It works by first generating a maze surrounded by walls without any structures by calling dfs and then by adding the ghost house and tunnel by calling fixMaze

Example

>>> generateMaze 15 15 342
>>> [[#,#,#,#,#,#,#,#,#,#,#,#,#,#,#],
     [#,.,.,.,.,.,.,.,#,#,.,.,.,.,#],
     [#,.,.,#,.,.,#,.,.,.,.,.,#,.,#],
     [#,.,#,#,.,.,.,.,.,.,#,.,.,.,#],
     [#,.,.,.,.,.,.,.,.,.,.,.,.,.,#],
     [#,., , , , , , , , , , , ,.,#],
     [#,#, ,#,#,#, , , ,#,#,#, ,.,#],
     [ ,., ,#, , , , , , , ,#, ,., ],
     [#,., ,#,#,#,#,#,#,#,#,#, ,.,#],
     [#,., , , , , , , , , , , ,#,#],
     [#,.,.,.,.,.,.,.,.,.,.,.,.,.,#],
     [#,#,.,.,.,.,.,.,.,.,.,#,.,.,#],
     [#,.,.,.,.,.,.,.,#,#,.,.,.,.,#],
     [#,.,.,.,#,.,.,.,.,.,.,.,.,.,#],
     [#,#,#,#,#,#,#,#,#,#,#,#,#,#,#]]