Tarefa2

Description

Module responsible for implementing the game logic.

Documentation

halfSpeed :: Double

The speed of ghost in Dead Mode (0.5)

Example

>>> halfSpeed
0.5

regularSpeed :: Double

The speed of ghost in Alive Mode (1.0)

Example

>>> regularSpeed
1.0

ghostSpawnChecks :: [Coords]

The places to check to spawn the ghosts, in relative position. (0,0) corresponds to the center of the ghost house, rounding to the left

getGhostSpawnCoords

Arguments

:: State
The given state

-> Coords
The coordinates of the next spawn position for a ghost

Returns the spawning position of the ghosts. This corresponds to the inside of the ghost house, from the center to the borders when the previous position is already taken, rounding to the left in case the width of the house is even

If there is no space available to spawn pacman, an error is thrown

getPacmanSpawnCoords

Arguments

:: State
The given state

-> Coords
The coordinates of pacman's spawn position

Returns pacman's spawning position. This corresponds to the position directly below the ghost house, rounding to the left in case the width of the house is even

If there is no space available to spawn pacman, an error is thrown

scareGhosts

Arguments

:: State
The original state

-> State
The resulting state

Changes the state of ghosts to Dead and reduces their speed to half

This function is called whenever pacman eats a food big

Example

>>> scareGhosts state
State {maze = [[#,#,#,#,#,#,#,#],[ ,*,.,.,.,M,W, ],[ ,.,.,<,.,#,., ],[#,#,#,#,#,#,#,#]], playersState = [Pacman (PacState {pacState = (0,(3,2),1.0,R,5,1),
timeMega = 0.0, openClosed = Open, pacmanMode = Normal}),Ghost (GhoState {ghostState = (1,(6,1),0.5,L,10,0), ghostMode = Dead})], level = 1}

respawnPlayers

Arguments

:: State
The original game state

-> State
The updated game state, with the players at their spawning positions

Resets the game state.

More specifically, it removes the players from their locations and places them at their spawning positions. It also revives dead ghosts and set pacman's time mega to 0.

This function is called when pacman loses a life and the game is reset.

pseudoMove

Arguments

:: Coords
The (y, x) size of the maze (used for the tunnel)

-> Orientation
The direction to move in

-> Coords
The initial position of the player

-> Coords
The coordinates of the square the player lands in

When combined with areAdjacent, defines the geometry of the board (currently the coordinates are mod n). Returns the coordinates of where a player lands should he move in the specified direction (Move does not have to be legal)

Examples

>>> pseudoMove (2,3) L
(2,2)

>>> pseudoMove (2,3) D
(3,3)

areAdjacent :: Maze -> Coords -> Coords -> Bool

When combined with pseudoMove, defines the geometry of the board (currently the coordinates are mod n). Returns whether two given cells of the maze are adjacent

Examples

>>> areAdjacent testMaze (2,3) (3,3)
True

>>> areAdjacent testMaze (2,2) (3,3)
False

move

Arguments

:: Int
The id of the player who will move

-> Coords
The new coordinates

-> State
The current state

-> State
The resulting state

This function moves pacman to the coords, sets his new position as Empty and calculates the resulting effects on the state:

If he walks into Empty, nothing happens

If he walks into a Food Little, he gains 1 point

If he walks into a Food Big, his mode becomes Mega, he gains 5 points and all ghosts become scared (their state becomes Dead and their speed is reduced to half)

If called by a non-pacman player, an error may be thrown

interactPacman

Arguments

:: [Player]
the list of players to interact with the main player

-> State
the original state

-> State
the resulting state

Computes the interaction player by player. This function is an auxiliar function to interation. If for some reason it is called by a non-Pacman player, an error may be thrown

interaction

Arguments

:: State
the original state

-> State
the resulting state

Calculates the result from the player interactions. These interactions happen when two players are at the same position:

A dead ghost is killed by pacman, which sends him back to the ghost house alive and causes pacman to gain 10 points

An alive ghost kills pacman, which resets the level and causes pacman to lose 1 life

If pacman had 0 lives when he interacts with an alive ghost, instead his state changes to Dying and all of the ghosts are removed from the game

Aditionally, if pacman runs out of timeMega, his state goes back to Normal and the ghosts' state to Alive. If called by a ghost, an error may be thrown

play

Arguments

:: Play
The specified play

-> State
The original game state

-> State
The resulting game state

Updates the game state according to the specified play:

If the player (assumed to be a pacman) is Dying, nothing happens

If the player tries to move in a diferent direction than his orientation, instead of moving he is rotated to the direction he tried to move towards

If the player should move (there is no Wall in front of him), this function calls move and then interaction

Example

>>> play (Move 0 L) state
State {maze = [[#,#,#,#,#,#,#,#],[ ,*,.,.,.,M,., ],[ ,.,.,>,.,#,., ],[#,#,#,#,#,#,#,#]], playersState = [Pacman (PacState {pacState = (0,(3,2),1.0,L,5,1),
timeMega = 0.0, openClosed = Open, pacmanMode = Normal}),Ghost (GhoState {ghostState = (1,(6,1),1.0,L,10,0), ghostMode = Alive})], level = 1}