Tarefa3

Description

Introdução

Foi pedido, após a geração automática de labirintos da Tarefa1, que se obtivesse uma forma de compactar o labirinto, aproveitando a repetição de elementos.

A solução para o problema que se pretende resolver é simples e não provocou dificuldade ou constrangimento.

Objetivo

O objetivo desta tarefa é comprimir um labirinto do Pacman, de forma a este poder ser representado através de uma menor quantidade de dados. Para isso, aproveita-se a repetição sucessiva de várias peças (nomeadamente paredes) nos labirintos.

Solução

Primeiro, começa-se por compactar todas as filas individualmente. Para isso, percorre-se cada fila, tendo em conta a peça encontrada anteriormente. Quando se encontra uma peça diferente, adiciona-se o par com essa peça e o número de vezes que aparece à fila compactada, repetindo o processo para o resto da fila, até esta ser vazia.

Finalmente, procuram-se por elementos repetidos no labirinto compactado. Para fazer isso, percorre-se a lista. Para cada elemento da lista, percorre-se a lista novamente, procurando por elementos iguais a este, substituindo-os pelo construtor Repeat.

Conclusão

Obteve-se assim um labirinto compactado. O grupo considera que esta tarefa não colocou qualquer dificuldade, sendo, por isso, a mais simples de todo o projeto.

Documentation

partitionCorridor

Arguments

:: Piece
The piece to compare the others against

-> Corridor
The given corridor

-> (Int, Corridor)
A tuple consisting of the number of consecutives elements in corridor equal to the given piece, and the remaining corridor

Partitions a corridor based on a given piece. It returns a tuple, representing the number of consecutive elements in the beggining of the list equal to the given piece, and the remaining list, i.e., the corridor starting at the first element different than the given piece. It works in the same way as the span function, only that the first element it returns is an Int corresponding to the length of the list instead of the actual list

Examples

>>> partitionCorridor Wall [Wall, Wall, Empty, Food Little, Food Little]
>>> (2,[ ,.,.])

>>> partitionCorridor Empty [Empty, Food Little, Wall, Empty]
>>> (1,[.,#, ])

compactCorridor

Arguments

:: Corridor
The uncompressed corridor

-> [(Int, Piece)]
The compressed corridor, as a list of (Int, Piece) and not an Instruction

Compacts a corridor to a list of '(Int, Piece)', but not into an Instruction

Examples

>>> compactCorridor [Wall, Wall, Empty, Food Little, Food Little]
>>> [(2,#),(1, ),(2,.)]

>>> compactCorridor [Empty, Empty, Food Little, Wall, Empty]
>>> [(2, ),(1,.),(1,#),(1, )]

removeEqualInstructions

Arguments

:: Instructions
The original set of instructions

-> Instructions
The compressed set of instructions

Replaces all identical instructions with a Repeat instruction

Examples

>>> removeEqualInstructions [Instruct [(1,Wall), (2, Empty)], Instruct [(3,Empty)], Instruct [(3,Empty)]]
>>> [Instruct [(1,#),(2, )],Instruct [(3, )],Repeat 1]

>>> removeEqualInstructions [Instruct [(2,Wall), (2, Empty)], Instruct [(3,Empty)], Instruct [(4,Empty)], Instruct [(2,Wall), (2, Empty)], Instruct [(2,Wall), (2,Empty)]]
>>> [Instruct [(2,#),(2, )],Instruct [(3, )],Instruct [(4, )],Repeat 0,Repeat 0]

removeEqualInstructionsAux

Arguments

:: Instructions
The original set of instructions

-> Int
The index of the current head of the list

-> Instructions
The new compressed set of instructions

Auxiliary function to removeEqualInstructions. This function uses an accumulator to know what is the index of the head of the list, in order to feed it to the removeInstructionsEqualTo, as it requires the index value

Example

>>> removeEqualInstructionsAux [Instruct [(1,Wall), (2, Empty)], Instruct [(3,Empty)], Instruct [(3,Empty)]] 0
>>> [Instruct [(1,#),(2, )],Instruct [(3, )],Repeat 1]

removeInstructionsEqualTo

Arguments

:: (Instruction, Int)
The instruction to compare the others against, and its respective index

-> Instructions
The original set of instructions

-> Instructions
The new set of instructions with all instances of the given instruction replaced with 'Repeat i'

Replaces all Instruction equal to the given Instruction with a Repeat Instruction

Example

>>> removeInstructionsEqualTo (Instruct [(3, Empty)], 0) [Instruct [(1,Wall), (2, Empty)], Instruct [(3,Empty)], Instruct [(3,Empty)]]
>>> [Instruct [(1,#),(2, )],Repeat 0,Repeat 0]

compactMaze

Arguments

:: Maze
The given maze

-> Instructions
The maze in instruction form

Compacts a maze into a set of instructions

Example

>>> compactMaze testMaze -- consider testMaze the same as defined in Tarefa2.hs
>>> [Instruct [(8,#)],Instruct [(1, ),(1,*),(3,.),(1,M),(1,.),(1, )],Instruct [(1, ),(2,.),(1,<),(1,.),(1,#),(1,.),(1, )],Repeat 0]