Tarefa5

Description

Module responsible for the behaviour of the ghosts.

Documentation

isGhostHere

Arguments

:: [Player]
the list of players (not necessarily ghosts)

-> Coords
the specified coordinates

-> Bool
wether any ghost from the list is at the coordinates

Given a list of players, this function returns wether some of the ghosts within that list are at the specified coordinates.

getGhostType

Arguments

:: [Player]
The list of players (ghosts or otherwise)

-> Int
The id of the specified ghost

-> Int
The type of the ghost (1 or 2)

Given the list of players, returns the type (1 or 2) of the specified ghost (given by id)

Type 1 is direct chase, where the ghost will follow the shortest path towards pacman

Type 2 is advanced chase, where the ghost follows the shortest path to the position 4 pieces ahead of pacman, and only then enters direct chase

getPacmanLocation

Arguments

:: [Player]
The list of players

-> Coords
The pacman's coordinates

Returns the Location of the pacman

It assumes there is only one pacman player, meaning it will return the first value it finds. It also supposes the player exists in the first place, meaning it is not defined if there is no pacman.

Example

>>> getPacmanLocation [(Pacman (PacState (0, (1,1), 2.0, U, 1,1) 2.0 Closed Mega)), (Ghost (GhoState (1,(2,1), 2.5, L,2,2) Dead))]
>>> (1,1)

getNeighbourDistances

Arguments

:: Coords
The given coordinates

-> Memoization
All the distances

-> Coords
The coordinates of the target

-> [Player]
The list of players

-> [Int]
The resulting list.

Given some coordinates, all the distances between every two points, and a target, returns the distances from the neighbouring cells and the cell itself to the target that do not contain a ghost, i.e, given a set of coordinates (y, x), it returns the distance from the squares (y - 1, x), (y + 1, x), (y, x - 1), (y, x + 1) and (y, x), in that order, as long as no ghost is contained in them.

Note

If any of the neighbouring cells is a wall (meaning it is unreachable) or has a ghost, that value is instead replaced by -1 in the final list.

filterIndices

Arguments

:: Int
The index of the first element

-> (a -> Bool)
The filtering condition

-> [a]
The given list

-> [Int]
The resulting list of indices

Given the index of the first element, a filter and a list, this function returns a list containing the indices of the elements that satisfy said filter (ordered in ascending order)

hash'

Arguments

:: State
the state

-> Int
the hash

Calculates the hash of a State by casting it to a String and then calculating its hash (Data.Hashable)

show'

Arguments

:: State
the state

-> String
the resulting string

Casts a State to a String. Used by hash'

rand'

Arguments

:: Int
n

-> State
the state

-> Int
the pseudo-random number from 0 to n

Given n, this function calculates a pseudo-random number from 0 to n using a state

getLongestDirection

Arguments

:: Coords
The coordinates of the given ghost

-> Memoization
The map representing the distances

-> Coords
The target the ghost is trying to escape

-> State
The state

-> Orientation
The optimal orientation for the ghost, i.e, the direction which escapes the pacman most optimally

Given the coordinates of a ghost, and a map representing the distances between every two points of the maze, returns which orientation should the ghost be in in order to escape the target most optimally, i.e, the furthest away from it If two or more maximum distances are found, a pseudo-random one based on the hash of the state is returned.

Note

This function is greedy. If caught in a dead end, the ghost will not try to escape since that would cause a local decrease in the distance to the target.

getShortestDirection

Arguments

:: Coords
The coordinates of the given ghost

-> Memoization
The map representing the distances

-> Coords
The coordinates of the target

-> State
The state

-> Orientation
The optimal orientation for the ghost, i.e, the direction which leads to the pacman the quickest

Given the coordinates of a ghost, a target, and a map representing the distances between every two points of the maze, returns which orientation should the ghost be in in order to get to the target in the shortest possible distance. If two or more minimum distances are found, a pseudo-random one based on the hash of the state is returned.

getAdvancedTarget

Arguments

:: Int
the specified number of moves

-> Maze
the maze where the player moves

-> Player
the player

-> Coords
the coordinates where he lands

For a certain player, returns the position he would land in were he to take the specified number of steps (assuming collisions with walls)

chaseMode

Arguments

:: State
The given state

-> Int
The ghost's id

-> Memoization
The Memoization object

-> Play
The ghost's move

Given a state, and the id of an Alive ghost, returns the its move

Note

It is not defined for Dead ghosts

scatterMode

Arguments

:: State
The given state

-> Int
The ghost's id

-> Memoization
The Memoization object

-> Play
The ghost's move

Given a state, and the id of a Dead ghost, returns the its move

Note

It is not defined for Alive ghosts

newst

Arguments

:: [Play]
The list of moves

-> State
The original state

-> State
The resulting state

Returns the state were the players to move in the specified directions

chooseMode

Arguments

:: State
The original state

-> Memoization
The Memoization object

-> [Play]
The list of moves already calculated

-> Player
The given player

-> [Play]
The resulting list of moves

For the given player, chooses the mode they are to move in and stores their move. The moves of previous players are taken into account to decide the moves of the current player.

ghostPlay'

Arguments

:: State
The given state

-> Memoization
The Memoization object

-> [Play]
The resulting list of moves

Given a state, and a Memoization, returns the list of moves for each ghost, in the order they appear in the players list

ghostPlay

Arguments

:: State
The given state

-> [Play]
The resulting list of moves

Given a state, returns the list of moves for each ghost, in the order they appear in the players list