Tarefa6

Description

Module responsible for the behaviour of the Pacman bot.

Documentation

type DPMap = Map Int (Orientation, Int)

Represents all the values already calculated by the program

defaultDepth :: Int

The default search depth

hash''

Arguments

:: State
the state

-> Int
the step

-> Int
the hash

Calculates the hash of a tuple (State, Int) by casting it to a String and then calculating its hash (Data.Hashable)

show''

Arguments

:: State
the state

-> Int
the step

-> String
the resulting string

Casts a tuple (State, Int) to a String. Used by hash''

rand''

Arguments

:: Int
n

-> Int
the step

-> State
the state

-> Int
the pseudo-random number from 0 to n

Given n, this function calculates a pseudo-random number from 0 to n using a step and a state

evaluate

Arguments

:: Memoization

-> State
The given state

-> Player
Pacman

-> Int
The number of steps since the begining of the game

-> Int
The evaluation of the state

The evaluation of any given state of the game.

This is the target function the bot tries to maximize.

maxValue'

Arguments

:: [(Orientation, Int)]
The given list

-> [(Orientation, Int)]
The elements with the largest Int component

Given a list of [(Orientation, Int)] returns the elements which have the biggest Int component.

iterateOverList

Arguments

:: (a -> c -> (b, c))
The given function

-> c
The original accumulator

-> [a]
The list of arguments

-> ([b], c)
The output of the functions: a list consisting of the first component of each function call, as well as the accumulator returned by the final call.

Given a function (always used with testOrientation), and a list of parameters, and one original accumulator, returns the output of the function (as a list of results) as well as the accumulator resulting from running the function through all said arguments

search

Arguments

:: Int
The given step

-> State
The given state

-> Memoization
The given Memoization table

-> [Int]
The hashes of the previously searched states

-> Int
The search depth

-> DPMap
The map used for dynamic programming

-> (Orientation, Int, DPMap)
The optimal orientation and evaluation, alongside the updated DP map

Given a certain state, runs through every possible move for Pacman (left, right, down and up), and returns not only the optimal orientation of the move, as well as the evaluation of the optimal state that can be reached, and the map used for dynamic programming purposes.

bot'

Arguments

:: Int
The given step

-> State
The given state

-> Memoization
The given Memoization object

-> Maybe Play
The optimal move

Given a step and step, as well as Memoization, returns the best move for pacman to play

bot

Arguments

:: Int
The given step

-> State
The given state

-> Maybe Play
The optimal move

Given a step and step, returns the best move for pacman to play