Types

Description

Module with all the datatypes definitions needed for the Pacman

Documentation

defaultDelayTime :: Integer

data State

Represents the state of the Game at a given time

Constructors

State

Fields

maze :: Maze

The maze of the game

playersState :: [Player]

The players in the game

level :: Int

The current level of the game

type Maze = [Corridor]

type Corridor = [Piece]

data Piece

Constructors

Food FoodType

PacPlayer Player

Empty

Wall

data Player

Represents a player. Each player can be uniquely described by a PacState or a GhoState depending if it is a pacman or a ghost

Constructors

Pacman PacState
Player is a pacman, described by its PacState

Ghost GhoState
Player is a ghost, described by its GhoState

data Orientation

Represents directions: Left Right Up / Down

Constructors

L
Left

R
Right

U
Up

D
Down

Null
Null

orientationsList :: [Orientation]

A list of all 5 orientations, arbitrarily ordered

data PacState

Represents the state of Pacman in a given point in time

Example

pac = PacState (0,(3,2),1,R,5,1) 0 Open Normal

Constructors

PacState

Fields

pacState :: PlayerState

timeMega :: Double

openClosed :: Mouth

pacmanMode :: PacMode

data GhoState

Represents the state of a ghost in a given point in time

Example

ghost = GhoState (0,(3,2),1,R,5,1) Alive

Constructors

GhoState

Fields

ghostState :: PlayerState

ghostMode :: GhostMode

type Coords = (Int, Int)

Coordinates in the maze, given as a (y, x) pair The maze can be as a two-dimensional grid, where the origin (0,0) is the top left corner. The x-axis is the horizontal axis and the y-axis the vertical axis. The y coordinate increases as you go down, and the x coordinate increases as you go to the right

type PlayerState = (Int, Coords, Double, Orientation, Int, Int)

Represents the state of a generic player (i.e. Pacman or ghosts) in a given point in time

data Mouth

Represents the state of Pacman's mouth (open or closed)

Constructors

Open

Closed

data PacMode

Represents the state of Pacman (Dying Normal or Mega)

Constructors

Dying

Mega

Normal

data GhostMode

Represents the state of a ghost (dead or alive)

Constructors

Dead

Alive

data FoodType

Represents the different kinds of food

Constructors

Big
Rarer, worth more points, Pacman enters Mega mode when it eats it

Little
More common, worth less, Pacman does not change mode when it eats it

data Color

Constructors

Blue

Green

Purple

Red

Yellow

None

data Play

Represents an action made by a player Each action is a movement of a player in a certain direction

Warning

An action with a certain orientation does not mean that the player will move in that direction, only that it will be facing in that direction after the action

Constructors

Move

Fields

Int

The id of the player who performs the action

Orientation

The orientation of the action

type Instructions = [Instruction]

Type used to compact the entire maze

data Instruction

Data type used to compact a corridor (horizontal)

Constructors

Instruct [(Int, Piece)]
There is the same piece in n consecutive squares

Repeat Int
The same as the instruction with index i

coloredString

Arguments

:: String
The given string

-> Color
The given color

-> String
The resulting string

Adds the given color to the given string so it can be displayed in the console.

The supported colors are red, blue, green, purple and yellow

placePlayersOnMap

Arguments

:: [Player]
The list of players

-> Maze
The maze with no players

-> Maze
The maze with the players

Places the players on map so it can be printed to the console

printMaze

Arguments

:: Maze
The given maze

-> String
The resulting string

Converts the maze to a string

printPlayerStats

Arguments

:: Player
The given player

-> String
The resulting string

Parses the PlayerState to a string

Example

>>> printPlayerStats (Pacman (PacState (0, (1,1), 1, R, 5, 1) 0 Open Normal))
"ID:0 Points:5 Lives:1\n"

getPlayerID

Arguments

:: Player
The given player

-> Int
The player's ID

Returns the player's ID

Example

>>> getPlayerID (Pacman (PacState (0, (1,1), 1, R, 5, 1) 0 Open Normal))
0

getPlayerPoints

Arguments

:: Player
The given player

-> Int
The player's points

Returns the player's points

Example

>>> getPlayerPoints (Pacman (PacState (0, (1,1), 1, R, 5, 1) 0 Open Normal))
5

setPlayerCoords

Arguments

:: Player
The given player

-> Coords
The new coordinates

-> Player
The resulting player

Changes the player's coordinates

Example

>>> setPlayerCoords (Pacman (PacState (0, (1,1), 1, R, 5, 1) 0 Open Normal)) (4,3)
Pacman (PacState (0, (4,3), 1, R, 5, 1) 0 Open Normal)

getPieceOrientation

Arguments

:: Piece
The given piece

-> Orientation
The orientation of the piece

Returns the orientation of the given piece

Example

>>> getPieceOrientation (PacPlayer (Pacman (PacState (0, (1,1), 1, R, 5, 1) 0 Open Normal)))
R

getPacmanMode

Arguments

:: Player
The given player

-> PacMode
Its PacMode

Returrns the pacMode of the player (only defined for Pacman)

Example

>>> getPieceOrientation (Pacman (PacState (0, (1,1), 1, R, 5, 1) 0 Open Normal))
Normal

getPlayerState

Arguments

:: Player
The given player

-> PlayerState
The player's PlayerState

Returns the PlayerState of a player

Example

>>> getPieceOrientation (Pacman (PacState (0, (1,1), 1, R, 5, 1) 0 Open Normal))
(PacState (0, (1,1), 1, R, 5, 1)

getPlayerOrientation

Arguments

:: Player
The given player

-> Orientation
The player's orientation

Returns the orientation of a player

Example

>>> getPlayerOrientation (Pacman (PacState (0, (1,1), 1, R, 5, 1) 0 Open Normal))
R

replaceElemInMaze

Arguments

:: Coords
The coordinates of the Piece to replace

-> Piece
The piece to replace the old one with

-> Maze
The original maze

-> Maze
The resulting maze

Replaces the element in the given coordinates with the given piece in a maze

replaceNElem

Arguments

:: Int
The position of the element

-> a
The element to replace the old one with

-> [a]
The original list

-> [a]
The resulting list

Replaces the n-th element in a list with a given element

getPlayerCoords

Arguments

:: Player
The given player

-> Coords
The coordinates of the player

Returns the coordinates of a given player

Example

>>> getPlayerCoords (Pacman (PacState (0, (1,1), 1, R, 5, 1) 0 Open Normal))
(1,1)