Tarefa1

Description

Introdução

Nesta tarefa desenvolvemos uma função que gera um labirinto válido para o jogo Pacman.

generateMaze :: Int -> Int -> Int -> Maze

Objetivo

O labirinto que queremos criar deve cumprir vários requisitos:

O labirinto é rodeado por peças Wall (exceto túnel);

A faixa mais central forma um túnel;

Todos os corredores têm o mesmo comprimento;

Existe uma estrutura denominada “casa dos fantasmas”, localizada no centro do labirinto, que respeita uma determinada estrutura

Para conseguirmos cumprir todos os requisitos necessários, começamos por obter um labirinto aleatório não válido, e,de seguida, substituimos todas as peças necessárias para o tornar válido.

Discussão e conclusão

Após testar a função de geração de labirintos válidos, verificámos que cumpre os requisitos listados

Documentation

generateMaze :: Int -> Int -> Int -> Maze

Dado o número de corredores (horizontais), o seu comprimento, e um número inteiro positivo (para usar como semente num gerador pseudo-aleatório) retorna um labirinto que cumpre os requisitos referidos anteriormente

geraAleatorios :: Int -> Int -> [Int]

Dada uma semente retorna uma lista de n inteiros gerados aleatoreamente

nrAleatorio :: Int -> Int

Dada uma semente retorna um inteiro gerado aleatoriamente

subLista :: Int -> [a] -> [[a]]

Transforma uma lista numa lista de listas de tamanho n

convertePeca :: Int -> Piece

Transforma um número inteiro numa Piece, tal que:

3 = Food Big

0 <= n < 70 = Food Little

70 <= n <= 9 = Wall

converteCorredor :: [Int] -> Corridor

Transforma uma lista de inteiros num corredor

converteLabirinto :: [[Int]] -> Maze

Transforma uma lista de inteiros num labirinto

poeParedes :: Maze -> Maze

Substitui todas peças que rodeiam o labirinto por Wall

tudoParedes :: Corridor -> Corridor

Substitui todas as peças de um corredor por Wall

Podia também ser definida da seguinte forma:

tudoParedes c = map (\p -> Wall) c

pEuParede :: Maze -> Maze

Substitui em todos os corredores a primeira e a última peça por Wall (exceto no último em que esta substituição é feita para todas as peças)

poeTuneis :: Int -> Maze -> Maze

Substitui, na faixa mais central, as extremidades por Empty, criando um túnel

Caso o número de linhas do labirinto for ímpar, o túnel é formado por um corredor, apenas

Caso o número de linhas do labirinto for par, o túnel é formado por dois corredor

poeTunelImpar :: Int -> Maze -> Maze

Substitui as extremidades por Empty, no corredor correspondente à linha n

poeTunelPar :: Int -> Maze -> Maze

Substitui as extremidades por Empty, nos corredores correspondentes à linha n e n+1

poeCasaFant :: Int -> Int -> Maze -> Maze

Coloca a casa dos fantasmas no labirinto e substitui as peças que a rodeiam por Empty

Propriedades

A casa dos fantasmas tem uma altura fixa igual a 3

Caso o número de colunas do labirinto seja ímpar, o comprimento da casa é 9

Caso o número de colunas do labirinto seja par, o comprimento da casa é 8

A casa está centralizada, execto nos casos em que não é possível (quando o número de linhas é par). Nestes casos a casa é colocada a X peças de distância da parede inferior e a X-1 peças de distância da parede superior

As peças que rodeiam a casa de fantasmas são Empty

cFantP :: Int -> Int -> Maze -> Maze

Coloca no labirinto a casa dos fantasmas (com comprimento igual a 9), apartir da linha n e da coluna c

colocarChaoP :: Int -> Corridor -> Corridor

Substitui as 10 peças após a posição x por Empty

Podia também ser definida sa seguinte forma:

colocarChaoP 0 l = replicate 10 Empty : drop 10 l
colocarChaoP x (h:t) = h: colocarChaoP (x-1) t

colocarEntradaP :: Int -> Corridor -> Corridor

Coloca a entrada da casa dos fantasmas (para as colunas pares)

colocarParedesP :: Int -> Corridor -> Corridor

Coloca as paredes laterais da casa dos fantasmas (para as colunas pares)

colocarBaseP :: Int -> Corridor -> Corridor

Coloca a base da casa dos fantasmas (para as colunas pares)

cFantI :: Int -> Int -> Maze -> Maze

Coloca no labirinto a casa dos fantasmas caso o número de colunas seja impar

colocarChaoI :: Int -> Corridor -> Corridor

Substitui as 11 peças após a posição x por Empty

colocarEntradaI :: Int -> Corridor -> Corridor

Coloca a entrada da casa dos fantasmas (para as colunas impares)

colocarParedesI :: Int -> Corridor -> Corridor

Coloca as paredes laterais da casa dos fantasmas (para as colunas impares)

colocarBaseI :: Int -> Corridor -> Corridor

Coloca a base da casa dos fantasmas (para as colunas impares)