Tarefa2

Description

Introdução

Nesta tarefa, definimos o resultado da movimentação de um jogador num estado de jogo

play :: Play -> State -> State

Objetivo

Para percebermos o efeito da jogada pretendida no jogo começamos por perceber se o jogador vai apenas mudar de direção ou se vai avançar. Se ele for avançar, procura-se ver se este irá ao encontro de outro jogador, aplicando as respetivas consequências se tal acontecer.

De seguida, move-se o jogador alterando o estado do jogo de acordo com a peça que se encontra no posição seguinte

Finalmente, verifica-se também se as comidas todas do labirinto já foram comidas, se for o caso, o nível muda. Consequentemente, o tempoMega do pacman diminui, a velocidade dos fantasmas no modo Alive aumenta e o labrinto é atualizado para um labirinto pré-definido.

Discussão e conclusão

Considerámos que a velocidade dos fantasmas à medida que os níveis passam aumenta 0.5, no entanto, para um máximo de 2, porque se não o pacman morriria muito rapidamente. Além disso, o tempoMega também diminui, mas para um mínimo de 5 segundos, para não tornar o jogo demasiado difícil.

Após testar os diferentes cenários possíveis, concluímos que o nossa função comporta-se como esperado.

Documentation

play :: Play -> State -> State

Move um jogador (identificado pelo seu identificador único (ID)) numa determinada direção e atualizando o efeito dessa jogada no labirinto atual

play_ :: Play -> Maze -> Player -> [Player] -> Int -> State

Verifica se um jogador se movimenta numa direção que não a sua direção atual, se for o caso a sua posição mantêm-se apenas se altera a orientação

Se não for o caso, descobre se existe algum/alguns jogadores na posição para onde se move. Se não existirem este apenas avança de posição, alterando possivelmente alguns dos seus atributos consoante a peça que se encontra nesta mesma posição Se existirem este altera os atributos dos jogadores que se encontram e, de seguida, faz algo semelhante ao caso anterior, ou seja, avança de posição e sofre algumas alterações consoante a peça que se encontra na posição para onde se move

separaPlayers :: Play -> [Player] -> (Player, [Player])

Separa o jogador que se vai mover dos restantes

isDying :: Player -> Bool

Vê se se trata de um pacman no modo Dying

playersInCoords :: Coords -> [Player] -> ([Player], [Player])

Divide os jogadores que se encontram numas certas coordenadas c dos restantes

Podia também ser definida sa seguinte forma:

playersInCoords c = (filter ((== c).(getPlayerCoords)) ,filter ((/= c).(getPlayerCoords)) )

mudaPlayers :: Maze -> Player -> [Player] -> Int -> (Player, [Player])

Altera o estado dos jogadores quando eles se cruzam, consoante o estado de cada um

Quando um Pacman encontra um ghost no modo Alive, pode acontecer uma de duas coisas. Se este ainda tiver vidas, perde uma vida . Se o Pacman já não tiver vidas, este altera o seu estado para Dying

Quando um Pacman encontra um ghost no modo Dead, este obtem +10 pontos, enquanto que o ghost desaparece, reeaparecendo dentro da casa dos fantasmas em modo Alive

vAlive :: Int -> Double

Retorna a velocidade com que o fantasma fica quando passa ao modo Alive, de acordo com o nível do jogo

proximoMovimento :: Play -> Maze -> Player -> [Player] -> Int -> Piece -> State

Retorna um State de acordo com o player dado e com a peça dada (peça que se encontra na posição para onde se move)

Se o jogador for um pacman:

Quando transita para uma posição vazia, nada acontece;

Quando transita para uma posição com comida pequena, a sua pontuação é atualizada com +1 ponto;

Quando transita para uma posição com comida grande, a sua pontuação é atualizada com +5 pontos, o seu estado muda para Mega e o timeMega muda para tMega. Além disso, os fantasmas existentes no labirinto mudam o seu estado para Dead e diminuem a sua velocidade para metade.

Quando a peça dada for uma parede, este não altera a sua posição

tempoML :: Int -> Double

Calcula o tempo que o Pacman fica no modo Mega após comer uma comida grande (que vai diminuindo com as passagens de nível)

avancaPosicao :: Play -> Maze -> Player -> [Player] -> Int -> State

Retorna um estado após avançar o jogador um casa (deixando a posição onde estava com a peça Empty)

changeMouth :: Player -> Player

Altera o estado da boca do Pacman

somaPontos :: Int -> Player -> Player

Soma x pontos ao jogador

ghostsDead :: [Player] -> [Player]

Altera o estado de todos os fantasmas para Dead e diminui a sua velocidade para metade

Obs: considera que todos os jogadores dados são fantasmas

inserePlayer :: Player -> [Player] -> [Player]

Insere um jogador numa lista e jogadores de forma a que na lsita resultante os IDs dos jogadores apareçam por ordem crescente

mudarDeNivel :: State -> State

Atualiza o estado do jogo para a mudança de nível, aumentando a velocidade dos fantasmas e atualizando o labirinto

comComida :: Maze -> Bool

Verifica se ainda existe alguma comida no labirinto

mudaVelocidadeP :: [Player] -> [Player]

Muda a velocidade dos fantasmas (quando se muda de nível) e a posição dos diferentes jogadores. O pacman vai para uma posição central, por baixo da casa dos fantasmas e os fantasmas aparecem numa das posições possíveis dentro da casa dos fantasmas, aleatoriamente.

mudaMaze :: Maze -> Maze

Retorna o labirinto do novo nível