Tarefa4

Description

Introdução

Nesta tarefa fazemos o estado do jogo atualizar para a passagem de um instante de tempo

Para tal é necessário definir a função passTime cujos inputs são um número real positivo, que representa o número de iterações por que o jogo já passou (step) e o estado atual do jogo. O resultado deverá ser um novo estado em que todos os jogadores e o mapa foram atualizados.

passTime :: Int  -> State -> State

Objetivo

Nesta atualização temos em conta três principais aspetos:

a velocidade de cada jogador

o tempoMega do Pacman

tipo de jogador

Velocidade : consoante a velocidade, cada jogador vai efetuar um diferente número de jogadas por iteração. Considerámos que a velocidade dos jogadores pode apenas variar entre 0.5 e 2, visto que com a adição da funcionaldiade da subida de nível, os fantasmas podem atingir, no máximo, uma velocidade de 2.

Tempo Mega : se o Pacman estiver no modo Mega, após a passagem de um instante de tempo este vai diminuir. Além disso, caso este chegue a 0, é necessário mudar tanto o modo do Pacman como o dos fantasmas para Normal e Alive, respetivamente.

Tipo de jogador : se o jogador for um Pacman, simplesmente movimenta-se no direção atual que tem. Se for um ghost, segue a estratégia usada na Tarefa 5 para determinar o seu movimento.

Discussão e conclusão

Após testar, verificámos que a função passTime avança todos os jogadores consoante o esperado.

Documentation

defaultDelayTime :: Integer

Tempo entre jogadas definido

passTime :: Int -> State -> State

passTime_ :: Int -> State -> [Player] -> State

Altera o estado do jogo, movendo todos os jogadores de acordo com a sua velocidade. Para poder controlar a velocidade é necessário perceber como a função se comportou na iteração anterior, para isso, a função recebe o step. Com esta informação é possível determinar quantas vezes o jogador deve progredir nesta iteração, utilizando a paridade deste número.

Assim, considera-se que:

se a velocidade for 0.5, o jogador apenas avança nas iterações com um step par;

se a velocidade for 1, o jogador avança uma vez por iteração;

se a velocidade for 1.5, o jogador avança uma vez nas iterações impares e duas nas pares;

se a velocidade for 2, o jogador avança duas vez por iteração.

joga :: Player -> State -> State

Dado um jogador e um estado de jogo retorna o estado resultante de efetuar uma jogada com esse mesmo jogador

tiraTempo :: State -> [Player] -> State

Caso o pacman esteja em modo Mega, retira-lhe o tempo equivalente a uma iteração, verificando se este volta ao modo Normal ao não. Caso contrário retorna o estado dado, sem alterações

mudaTempo :: [Player] -> [Player]

Retira ao timeMega do Pacman o tempo equivalente a uma iteração

mudaEstado :: [Player] -> Int -> [Player]

Altera todos os jogadores sabendo que o pacman irá mudar do estado Mega para Normal. Ou seja, os fantasmas ficam todos no modo Alive e o Pacman além de ficar no estado Normal, altera o seu timeMega para 0