Tarefa5

Description

Introdução

A tarefa 5 centra-se a empreender o comportamento dos fantasmas.

Objetivo

O objetivo desta tarefa é definir a função ghostPlay cujo input é o Estado dos Fantasmas. O resultado deverá ser uma lista de jogadas que o fantasma realizará de modo a atingir o seu objetivo, dependendo do seu Estado (Dead ou Alive).

ghostPlay :: State -> [Play]

Teremos então duas hipóteses de ação dos fantasmas:

chaseMode;

scatterMode.

chaseMode : quando os fantasmas estão Alive e o Pacman está Normal, os fantasmas irão perseguir o Pacman.

scatterMode : quando os fantasmas estão Dead e o Pacman em modo Mega, os fantasmas terão que fugir do Pacman.

Discussão e Conclusão:

Após várias tentivas de funções, decidimos que a função que usa uma lista de coordenadas e orientação para detereminar o melhor caminho até às coordenadas do Pacman naquele momento seria a mais eficiente pois evita que o fantasma fique preso em becos. De seguida testamos e confirmamos a eficiência desta função.

Documentation

ghostPlay :: State -> [Play]

semPacman :: [Player] -> [Player]

Lista de Jogadores sem o Pacman.

veFantasmas :: State -> Player -> Play

Prevê o estado que o fantasma terá na jogada seguinte e de acordo com o estado aplica uma de duas funções.

chaseMode :: State -> Int -> Play

Dependendo do estado do fantasma naquele momento, temos duas hipóteses de comportamento dos fantasmas:

Quando um fantasma está em modo Alive, vamos aplicar o chaseMode. O fantasma irá perseguir o Pacman, determoinando a posição deste último sempre que bate numa parede ou o Pacman ou mude de Estado (Alive -> Dead).

chaseMode :: State -> Int -> Play

scatterMode :: State -> Int -> Play

Quando um fantasma está em modo Dead, toma o comportamento scatterMode Sempre que o fantasma se depara com uma parede, irá mover-se sempre para a sua direita até voltar a Alive.

scatterMode :: State -> Int -> Play

encontraPacman :: [Player] -> Coords

Procura numa lista de jogadores as coordenadas do pacman

bestWay :: Maze -> Coords -> [(Coords, Orientation)] -> Orientation -> Orientation

Escolhe o melhor caminho tendo em conta o labirinto, as coordenadas do Pacman e uma lista que contém todas possibilidades de movimento dos fantasmas. De modo a escolher o caminho que melhor se adequa, ao percorrer a lista de possibilidades de jogadas, vamos mudar a peça para Paredes nos locais onde o ghost se encontra, impedindo que o Pacman volte para trás ou fique preso no labirinto.

temParede :: Maze -> Coords -> Orientation -> [(Coords, Orientation)]

Verifica se tem parede ou não nas posições à volta do fantasma

possibilidadesMov :: Maze -> Coords -> Orientation -> [(Coords, Orientation)]

Aplica a função temParede a todas as orientações.

mudarOrient :: Orientation -> [(Coords, Orientation)] -> [(Coords, Orientation)]

Para o ghost encontrar o pacman, primeiramente, terá que tomar a orientação Null e depois orientação que é a primeira da lista