Tarefa6

Description

Introdução

Nesta tarefa implementamos um robô/bot que joga Pacman automaticamente. Este, a partir do inteiro e do estado do jogo que recebe, analisa o que o rodeia e procura tomar a melhor decisão.

Objetivos

O objetivo do nosso bot é tentar obter a maior pontuação possível. Para tal, definimos uma estratégia para cada situação do jogo.

Consideramos que quando o Pacman se encontra no modo Normal, este apenas se vai dirigir para a comida mais próxima se não existir nenhum fantasma por perto (cujo caminho mais curto tem uma distância inferior a 5). Se existir, este afasta-se dele.

Quando o Pacman está no modo Mega este começa por procurar fantasmas no modo Dead ao seu redor (que este consiga atingir com o tempo Mega que lhe resta e cuja distância é menor que 5). Se encontrar algum dirige-se para ele, caso contrário tem um comportamento idêntico ao Pacman no modo Normal.

Discussão e conclusão

Após testar a eficácia do bot percebemos que ele inicialmente toma boas decisões, mas à medida que o jogo vai avançando este começa a ter menos opções, porque a comida do labirinto já é limitada, e começa a ter mais dificuldades.

Documentation

bot :: Int -> State -> Maybe Play

Em cada instante, o robô tem apenas conhecimento do identificador do jogador a movimentar e do estado atual do jogo. Com esta informação, o robô vai decidir se quer efetuar alguma jogada ou não, consoante a estratégia delineada.

jogaOuNao :: Orientation -> Play -> Maybe Play

Verifica se a jogada p encontrado como sendo a ideal altera o comportamento normal do Pacman (mover-se na direção atual). Devolvendo Nothing caso não altere nada e caso contrário devolve Just p

jogadaBot :: Player -> State -> Play

Procura o jogada ideal para o jogador p dado, utilizando diferentes estratégias consoante o seu estado

botNormal :: Player -> State -> Play

Com o objetivo de retornar a melhor jogada que o Pacman pode fazer quando está no modo Normal, primeiro verifica-se se existe algum fantasma cujo comprimento de caminho mais curto seja menor que a distGhost (definida nos Types). Se existirem fantasmas ele procura a melhor forma de se afastar deles. Se não existirem o Pacman apenas procura dirigir-se para a comida mais próxima.

ghostsMaisProx :: [(Coords, Double, Orientation)] -> Maze -> [Player] -> Orientation -> [(Double, Orientation)]

Retorna a lista dos pares com as orientações e o tamanho do caminho que levam o Pacman a chegar a um fantasma (Alive). Para as descobrir é usado o algoritmo Breadth-First, ou seja, a partir de uma posição inicial vai se vendo quais os sítios que se pode alcançar até chegar ao objetivo, neste caso o objetivo é a distancia ultrapassar a distGhost

temGhostAlive :: Coords -> [Player] -> Bool

Dada uma posição c e uma lista de jogadores, verifica se em c existe algum fantasma no modo Alive

temGhost :: Coords -> [Player] -> Bool

Dada uma posição c e uma lista de jogadores, verifica se em c existe algum fantasma

mudaPossib :: Double -> Orientation -> [(Coords, Orientation)] -> [(Coords, Double, Orientation)]

Dado um número x, acrescenta a cada par esse número (tranformando o par num triplo. Além disso, consoante a orientação dada, muda ou não as orientações dos elementos da lista

comidaMaisProx :: [(Coords, Orientation)] -> Maze -> Orientation -> Orientation

Com o objetivo de descobrir para que direção o pacman deve ir para se aproximar de uma comida, é utilizado novamente o algoritmo Breadth-First

decideOrientacao :: [(Double, Orientation)] -> Coords -> Maze -> Orientation -> Orientation

Recebe uma lista com as orientações que levarão o Pacman a encontrar um fantasma, com a respetiva distância a este. Com esta informação, procura a melhor forma de evitar os fantasmas. Se dentro das possibilidades de movimento existirem orientações que não estão na lista dada, é escolhida uma delas aleatoriamente. Caso contrário, escolhe-se aquela cujo fantasma está mais longe

Obs: Decidimos que a orientação escolhida, quando existe possibilidade de escolha, é aleatória visto que quando experimentámos escolhendo a primeira, o pacman, numa dada altura do jogo limitava-se a fugir dos fantasmas, dando voltas ao labirinto até eventualmente morrer

getRandomElem :: [Orientation] -> IO Orientation

orientDiferentes :: [(Coords, Orientation)] -> [(Double, Orientation)] -> [Orientation]

Retorna todos as orientações que aparecem na primeira lista, mas não aparecem na segunda

maisLonge :: [(Double, Orientation)] -> (Double, Orientation) -> Orientation

Retorna a orientação cujo o número correspondente é o maior

botMega :: Player -> State -> Play

Com o objetivo de retornar a melhor jogada que o Pacman pode fazer quando está no modo Mega, primeiro verifica-se se existe algum fantasma Dead cujo comprimento de caminho mais curto seja atingivel no tempo que resta no modo Mega e menor que a distGhost (definida nos Types). Se existir ele procura aproximar-se dele, se não terá um comportamento semelhante ao pacman quando está no modo Normal.

getMoves :: Player -> Double

Descobre o número de máximo de jogadas que queremos analisar consoante o tempo Mega que o pacman ainda tem. Se der para fazer um número de jogadas igual ou superior à distGhost, paenas devolve a distGhost. Se não devolve o número de jogadas que consegue efetuar

ghostMaisProx :: [(Coords, Double, Orientation)] -> Maze -> [Player] -> Double -> Orientation -> Maybe Orientation

Caso exista um fantasma no modo Dead alcançavel com um número restrito de jogadas (moves) devolve o orientação que leva o Pacman a este. Caso contrário, devolve Nothing.